refact: move TokenInfo and TokenInfoVersion classes to token_info.py
